feat: make tests pass on Gemini 1.5 Flash (almost)
There is 1 failing test ("Uncomplete a task") that passes if the system
prompt is removed for ActorAgent. It is important to investigate this
further because the current system prompt has too much of an impact on
the model behavior and it would be great to avoid it if possible.
A simple attempt to remove it causes failures with OpenAI so let's
postpone it for now.