feat: refine prompts to make tests pass on all models
In order to use the fastest/cheapest models from providers, we need to
be more precise in instructions (e.g. "mark task as completed" vs
"complete task"). This proves to show better results for Gemini and Claude.

Notably, all tests except for drag&drop now pass on the faster Anthropic
model (Claude 3 Haiku). The only failure seems to be related to the fact
that Haiku prefers to spatially reason based on ARIA tree, not screenshot.
We can further improve it tweaking prompts, but for now it's good enough.