feat: implement basic caching for actions
This allows to skip LLM communication when acting on the same ARIA tree.
The assumption is that as long as the goal and ARIA are the same, we can
used the cached tools from LLM and execute this right away.

It only works for in-memory cache and doesn't handle incorrect responses
from LLM at the moment.