feat!: visionless verifications
Turns out we can skip vision in most cases where the screenshot is not really needed (e.g. check for the presence of some elements or their state). This reduces cost (e.g. OpenAI screenshot is 25k tokens regardless of the size) and improves performance.

We can enable screenshots and vision only in tests that require spatial reasoning such as a drag and drop test.