revert: retry actions
The retry serves as a simplistic manager agent. In a situation when the
model attempts to do too much on the page. For example, in Google search
it's common to for model to suggest to type word in the search box,
press Enter and then click "Search" button. The last step is redundant
and causes the test to fail. Retry would attempt to search again and the
model would respond that nothing needs to be done since the goal is
achieved.

Eventually, we want to replace retry with a more intelligent agent that
would constantly evaluate what needs to be done to achieve the goal as
the action is executed.

This reverts commits e1b734d, 961d35a and bb6b59d.