feat: retry verification if page is loading
This is a basic retry implementation that catches verification errors
and checks whether the page is loading. If it does, it would wait 0.5s
and retry up to 5s.

VerifierAgent should not really know about LoadingDetectorAgent,
a high-level abstraction is needed here (e.g. CoordinatorAgent).
It's worth to add it later when we need to add retries for ActorAgent too.