ci: create release/tag on `release:` commit
This simplifies release automation by removing the need to create a tag
and run two CI workflows - one for commit and one for the tag.

Instead, simply bump the version in `pyproject.toml` and commit with a
message starting with `release:`. The CI workflow will read a version
from `pyproject.toml`, build a package, and create a release and a tag
on GitHub.