feat: per-test LLM caching
This commit adds LLM caching support in Alumnium implemented on
top of SQLite to simplify cache saving/restoring on CI.

The overall idea for the test is the following:
1. Cache is enabled by default for all LLM communications.
2. It automatically starts tracking the requests to LLM and saves
   them to a database transaction. This transaction effectively
   serves as an in-memory cache allowing to reuse of the cache
   during a test execution.
3. Once the test is completed, the user should do one of the following:
  a. _save_ the cache gathered during the test
     _if the test successfully passes_, or
  b. _discard_ the cache _if the test failed_.
4. Saving the cache would commit the transaction to the
   database file so subsequent/parallel tests can use it.
5. Discard the cache would roll back the transaction so
   subsequent/parallel tests would attempt to communicate with LLMs again.

Effectively, a test serves as an oracle.
A passing test indicates that Alumnium performed do/check/get correctly.
A failing test indicates the opposite - something was wrong during
execution so it doesn't make sense to persist LLM choices.

There is still a lot to be done with the whole caching
(e.g. cache max size, FIFO or LRU, etc.), so this is just a foundation.

Fixes #59.