chore: update dependencies and configuration for ruff
The poetry.lock lockfile was also updated.

The ruff package was added and the flake8, flake8-pyproject, black, isort and autoflake packages were removed from the poetry project.

Ruff was configured to have similar behavior, including implementing linting for the E, F, W rules, same ignored directories, same ignored rules.

Some comments:
- ruff does not implement W503. It was removed from the ignore list
https://github.com/astral-sh/ruff/issues/4125

- From what I see, isort is implemented in ruff, but the
isort.profile config is not available
https://docs.astral.sh/ruff/settings/#lintisort

- To have equivalent setup to the profile, you could use these configs in isort
https://pycqa.github.io/isort/docs/configuration/profiles.html
The problem with this is that only one of those configs is available in [tool.ruff.isort]. I am not sure if the behavior will be equivalent.
I assume that everything will work as ruff was made with the intention of replacing
Black and includes the same features.

- Another page that mentions black and isort
https://pycqa.github.io/isort/docs/configuration/black_compatibility.html

- ruff mentions implementing autoflake, but there are no references to
autoflake in the settings page
https://docs.astral.sh/ruff/faq/#how-does-ruffs-formatter-compare-to-black
https://docs.astral.sh/ruff/settings