ci: run tests on PRs from forks
The current CI setup doesn't allow running tests on PRs from forks
because the secrets are not available. The suggested workaround is
described in https://michaelheap.com/access-secrets-from-forks/:

1. Use `pull_request_target` event instead of `pull_request`.
2. Check if PR is from a user who has write access to the repo.
3. If not, fail the job.
4. If yes, run the job with the secrets.

The logic is that when PR is opened from a fork, the workflow is
going to fail. Then, when the maintainer reviews the PR, they can
re-run the failed workflow and this time it will have access to the
secrets

There is a security risk that something in PR could steal tokens
(e.g. dependency in poetry.lock) so extra attention from maintainers is
required. However, since all AI providers are capped with a small budget,
the overall risk is negligible.