feat!: upgrade from haiku 3 to haiku 4.5
Switches from Haiku 3 to Haiku 4.5. We skipped Haiku 3.5 because it
never worked really well for us, but it looks like Haiku 3 will be
phased out eventually.

With a new model, we can finally adopt a structured planner similar to
the one used in OpenAI, Google, etc. This improves accuracy as it gives
an output token space to produce a reasoning first. The next step would
be to adopt a thinking mode that this version of Haiku supports.

It's worth calling out that:
- Bedrock rate limits are really low (I hit just after a few requests)
- Bedrock doesn't mention that the model supports vision,
  even though it works just fine
- price is 4x higher, with $1 vs $0.25 per MToK.

Fixes #106