fix: follow open url instructions
Right now, the planner is not correctly instructed
for the variety of tools that the actor supports.
Notably, when Alumni is started with an extra
NavigateToUrlTool, it is not being selected because
the planner prefers to look at the accessibility
tree links instead of opening the URL directly,
which doesn't even need the tree at all.
It often results in incorrect tool selection,
so Alumnium cannot open the URL.

Until #109 is solved and we have a single actor
and planner, it makes sense to propagate the exact
tool names and specific tool examples to the
planner system prompt.