feat: automatically switch to new tab
This commit changes Playwright and Selenium to automatically switch to a new opened tabs during the test execution. This is useful in most cases because when user clicks a link which opens a new tab, they usually want to switch to it and continue with their test. Ideally, Alumnium would be able to switch between tabs on demand (e.g. al.do("switch to a new tab")) but this workaround should cover most cases for now.